move functions from sbi.h to sbicall.c file
