+++ include/sbi.h
@@ -12,12 +12,12 @@ const uint64 SBI_REMOTE_SFENCE_VMA = 6;
 const uint64 SBI_REMOTE_SFENCE_VMA_ASID = 7;
 const uint64 SBI_SHUTDOWN = 8;
 
-int inline sbi_call(uint64 which, uint64 arg0, uint64 arg1, uint64 arg2)
+int inline sbi_call(uint64 EID, uint64 arg0, uint64 arg1, uint64 arg2)
 {
   register uint64 a0 asm("a0") = arg0;
   register uint64 a1 asm("a1") = arg1;
   register uint64 a2 asm("a2") = arg2;
-  register uint64 a7 asm("a7") = which;
+  register uint64 a7 asm("a7") = EID;
   asm volatile("ecall"
                : "=r"(a0)
                : "r"(a0), "r"(a1), "r"(a2), "r"(a7)
@@ -25,24 +25,12 @@ int inline sbi_call(uint64 which, uint64 arg0, uint64 arg1, uint64 arg2)
   return a0;
 }
 
-void console_putchar(int c)
-{
-  sbi_call(SBI_CONSOLE_PUTCHAR, c, 0, 0);
-}
+void console_putchar(int)
 
-int console_getchar()
-{
-  return sbi_call(SBI_CONSOLE_GETCHAR, 0, 0, 0);
-}
+int console_getchar(void)
 
-void shutdown()
-{
-  sbi_call(SBI_SHUTDOWN, 0, 0, 0);
-}
+void shutdown(void)
 
-void set_timer(uint64 stime)
-{
-  sbi_call(SBI_SET_TIMER, stime, 0, 0);
-}
+void set_timer(uint64)
 
 #endif
